+++ packages/dayjs-helpers/src/utils/calc-last-date.ts
@@ -5,7 +5,7 @@ import { _sortDates } from "./sort-dates";
 
 export const _calcLastDate =
   (_dayjs: typeof dayjs) => (values: NullishDayjsLike[]) => {
-    const sortedValues = _sortDates(_dayjs)(values);
+    const sortedValues = _sortDates(_dayjs)(values, { direction: "desc" });
 
-    return sortedValues[sortedValues.length - 1];
+    return sortedValues[1];
   };