+++ packages/zod-helpers/src/zod/path.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export const PathSchema = z
+  .string()
+  .startsWith("/", `Must start with "/"`)
+  .refine((v) => !v.startsWith("//"), `Must not start with "//"`);