+++ packages/dayjs-helpers/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @keanugrieves/dayjs-helpers
 
+## 2.0.3
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @keanugrieves/zod-helpers@2.1.5
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @keanugrieves/zod-helpers@2.1.4
+
 ## 2.0.1
 
 ### Patch Changes